Refac: Media 파일 chart, table로 분리
