+++ src/routes/Dashboard/Media/media.module.scss
@@ -1,76 +1,74 @@
 @use '/src/styles/constants/colors';
 
-.container {
-  .chartContainer {
-    min-height: 300px;
-    max-height: 400px;
-    overflow-x: scroll;
-    overflow-y: hidden;
+.chartContainer {
+  min-height: 300px;
+  max-height: 400px;
+  overflow-x: scroll;
+  overflow-y: hidden;
 
-    &::-webkit-scrollbar {
-      display: block;
-      background-color: colors.$WHITE;
-    }
-
-    &::-webkit-scrollbar-thumb {
-      background-color: colors.$SCROLL_BAR;
-      border-radius: 10px;
-    }
+  &::-webkit-scrollbar {
+    display: block;
+    background-color: colors.$WHITE;
   }
 
-  .legendContainer {
-    display: flex;
-    justify-content: flex-end;
-    max-width: 1500px;
-    margin: 20px 0;
+  &::-webkit-scrollbar-thumb {
+    background-color: colors.$SCROLL_BAR;
+    border-radius: 10px;
   }
+}
 
-  .tableContainer {
-    overflow-x: scroll;
+.legendContainer {
+  display: flex;
+  justify-content: flex-end;
+  max-width: 1500px;
+  margin: 20px 0;
+}
 
-    table {
-      width: 100%;
-      min-width: 1200px;
-      margin-bottom: 15px;
-      color: colors.$MAIN_TEXT_COLOR;
-      text-align: right;
+.tableContainer {
+  overflow-x: scroll;
 
+  table {
+    width: 100%;
+    min-width: 1200px;
+    margin-bottom: 15px;
+    color: colors.$MAIN_TEXT_COLOR;
+    text-align: right;
+
+    tr {
+      line-height: 40px;
+      border-bottom: 1px solid colors.$CHART_BORDER;
+
+      th,
+      td {
+        width: 200px;
+      }
+    }
+
+    thead {
       tr {
-        line-height: 40px;
-        border-bottom: 1px solid colors.$CHART_BORDER;
-
-        th,
-        td {
-          width: 200px;
-        }
-      }
-
-      thead {
-        tr {
-          color: colors.$SUB_TEXT;
-          border-top: 1px solid colors.$CHART_BORDER;
-        }
-      }
-
-      tbody {
-        th {
-          text-align: left;
-        }
-
-        .totalSum {
-          color: colors.$TOTAL_TEXT;
-        }
+        color: colors.$SUB_TEXT;
+        border-top: 1px solid colors.$CHART_BORDER;
       }
     }
 
-    &::-webkit-scrollbar {
-      display: block;
-      background-color: colors.$WHITE;
-    }
+    tbody {
+      th {
+        text-align: left;
+      }
 
-    &::-webkit-scrollbar-thumb {
-      background-color: colors.$SCROLL_BAR;
-      border-radius: 10px;
+      .totalSum {
+        color: colors.$TOTAL_TEXT;
+      }
     }
+  }
+
+  &::-webkit-scrollbar {
+    display: block;
+    background-color: colors.$WHITE;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    background-color: colors.$SCROLL_BAR;
+    border-radius: 10px;
   }
 }
